Finish brace function
Finished writing the brace function. It has a chance to reflect
the fish's damage back at it as an option for players
to roll the dice and save bait. Will probably need further balancing.

Also updated reel function to only cost bait and have a chance of breaking rod
if fishHealth is above 0. This is to stop the game from using bait to reel in and exhausted fish.