Fix rewards for sea fish
The logic for detecting if a fish is a sea fish wasn't working properly
because it was set to check if the fish was caught in the open seas,
but the code already updates to sea battle by that point, so it was always false.
Changed "open seas" to "sea battle" in check to solve the problem.