Add reel cooldown and new hook item
The reel now has a random cooldown of 1 to 5 and players have to use brace in between.
This creates a more dynamic battle instead of just spamming reel.
Added a hook item that acts as level gates that players need to purchase in the store.
This adds more playibilty as players now have something to work towards.